Adds summarized energy file.
This file was created by running the python measurements module/script on each of the separate energy.sqlite databases that exist for each test result. The entire database was then dumped to sql. All of those eight sql files were combined to form this file. All tables other than energy were dropped. This leaves just the overall energy summary for all of the tests in a single table, in this database file.